Added AROS specific macros to make it work with AROS 68K
Added a small change to the menuitem disabling code which is a workaround
for a problem when using Zune. Zune will rebuild the entire menu structure
each time a menuitem is disabled, so by making sure we only change the state
for menuitems that actually change state, the performance increases quite
considerably.